Move trim_ws from masternodes/token to strencodings
