+++ src/masternodes/tokens.cpp
@@ -8,24 +8,13 @@
 #include <chainparams.h> // Params()
 #include <core_io.h>
 #include <primitives/transaction.h>
+#include <util/strencodings.h>
 
 #include <univalue.h>
 
 const DCT_ID CTokensView::DCT_ID_START = DCT_ID{128};
 
 extern const std::string CURRENCY_UNIT;
-
-std::string trim_ws(std::string const & str)
-{
-    std::string const ws = " \n\r\t";
-    size_t first = str.find_first_not_of(ws);
-    if (std::string::npos == first)
-    {
-        return str;
-    }
-    size_t last = str.find_last_not_of(ws);
-    return str.substr(first, (last - first + 1));
-}
 
 std::optional<CTokensView::CTokenImpl> CTokensView::GetToken(DCT_ID id) const
 {
