+++ src/masternodes/tokens.h
@@ -16,8 +16,6 @@
 class CTransaction;
 class UniValue;
 
-std::string trim_ws(std::string const & str);
-
 class CToken
 {
 public:
